A project my by Eclipse
This application connects to Goeurp API to get countries list. The
application gets the current location by the GPS and asks to turn it on
if it is switched off. The HTTPs request return a JSON array that will
be sorted by distance to the current location the displayed in the
spinners. The use can choose his language that will be transmitted in
the request to get the suitable countries list. The current date is
displayed by default and a imagebutton allows the user to choose another
date